ProtectedGetX method replicates previous behaviour of GetX for ROOT6
Since ROOT6, the TF1::GetX() method can no longer be used without precaution.

In ROOT5, if the value val was larger or smaller than the maximum or minimum value of
the function (within its defined range) then TF1::GetX() would return X of the
maximum or minimum, respectively.

In ROOT6, TF1::GetX() has been reimplemented so that, in the above case, it
prints out 3 warning messages and returns a TMath::QuietNaN()! Therefore we provide this
function in order to replicate the previous behaviour.

The status = 0 if the value is within the limits of the function.
status = +1 if value is above the maximum, or status = -1 if value is below the minimum

When compiled with ROOT5, this method just calls TF1::GetX() (and status = 0 always).